refactor: move (( expression )) from test_command to compound_statement
With the introduced changes, some of the examples were failing with a
specific pattern, namely the files:

- `examples/wild-corpus/esoteric/lishp/evaluator.sh`
- `examples/wild-corpus/esoteric/lishp/test.sh`
- `examples/wild-corpus/esoteric/lishp/variables.map.sh`
- `examples/wild-corpus/esoteric/lishp/variables.sh`

When taking a closer look it was noted that the following program was
failing to parse:

```
(( size=1, max_index=10 )
```

After taking a closer look at bash's manual, it seems this should be
possible and it should be considered a compound command instead. As
such, this commit moves this pattern from `test_command` to
`compound_statement`, updating it so that we can have multiple
expressions separated by a comma.

Note that these changes will very very very likely break the existing
tests, as we recently updated some to expect `test_command`. However,
for the time being, I'm keeping those failures since I'd like to have
someone else take another look at the changes I'm introducing with this
commit to see if they make sense.