test: commit updated good tests
Add the updated tests for both `test/corpus/literals.txt` and
`test/corpus/statements.txt`, while leaving the failure in
`test/corpus/commands.txt`, as the two first simply updated the
`test_command` node to `compound_statement`, which is expected.

The latter broke the binary expression in a parenthesized expression in
a subscript, which is simply returning word instead of actually parsing
the binary or unary expression, still need to figure that out.